[Garima] Refactor Login Screen and LandingScreen for styling and structure. Correct the folder naming structure. Remove unused packages
