+++ package.json
@@ -46,13 +46,11 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
-    "react-native-extended-stylesheet": "^0.12.0",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-modal": "^13.0.0",
     "react-native-reanimated": "^2.2.4",
     "react-native-safe-area-context": "^3.3.2",
     "react-native-svg": "^12.1.1",
-    "react-native-svg-transformer": "^0.14.3",
     "react-native-web": "0.17.1",
     "react-navigation": "^4.4.4",
     "react-navigation-stack": "^2.10.4",
@@ -79,7 +77,6 @@
     "lint-staged": "^12.1.2",
     "lodash": "^4.17.21",
     "prettier": "^2.4.1",
-    "react-hook-form": "^7.20.2",
     "tslint": "^6.1.3",
     "typescript": "^4.3.5"
   },
